moved typedef check to atlasGraphUtils
