+++ .github/workflows/maven.yml
@@ -37,7 +37,6 @@ on:
       - ms-864-keycloak-jwks-internal-url
       - ms-610-ranges-policy-cache
       - ms-901-hidelineage
-      - ms-894-override-classification
     paths-ignore:
       - '.claude/**'
       - '.cursor/**'
