+++ repository/src/main/java/org/apache/atlas/repository/store/graph/v2/AtlasClassificationDefStoreV2.java
@@ -306,7 +306,7 @@ class AtlasClassificationDefStoreV2 extends AtlasAbstractDefStoreV2<AtlasClassif
 
         AtlasVertex ret = typeDefStore.findTypeVertexByNameAndCategory(name, TypeCategory.TRAIT);
 
-        if (AtlasGraphUtilsV2.classificationHasReferences(name)) {
+        if (AtlasTypeDefESUtils.classificationHasReferences(name)) {
             String displayName = "";
             if (existingDef != null) {
                 displayName = existingDef.getDisplayName() != null ? existingDef.getDisplayName() : "";
@@ -341,7 +341,7 @@ class AtlasClassificationDefStoreV2 extends AtlasAbstractDefStoreV2<AtlasClassif
 
         String typeName = AtlasGraphUtilsV2.getEncodedProperty(ret, TYPENAME_PROPERTY_KEY, String.class);
 
-        if (AtlasGraphUtilsV2.classificationHasReferences(typeName)) {
+        if (AtlasTypeDefESUtils.classificationHasReferences(typeName)) {
             String displayName = "";
             if (existingDef != null) {
                 displayName = existingDef.getDisplayName() != null ? existingDef.getDisplayName() : "";
